+++ gradle.properties
@@ -28,7 +28,7 @@ loader_version_range=[4,)
 
 # The unique mod identifier for the mod. Must be lowercase in English locale. Must fit the regex [a-z][a-z0-9_]{1,63}
 # Must match the String constant located in the main mod class annotated with @Mod.
-mod_id=runicrepository
+mod_id=runic_repository
 # The human-readable display name for the mod.
 mod_name=Runic Repository
 # The license of the mod. Review your options at https://choosealicense.com/. All Rights Reserved is the default.
@@ -38,7 +38,7 @@ mod_version=0.1.1
 # The group ID for the mod. It is only important when publishing as an artifact to a Maven repository.
 # This should match the base package used for the mod sources.
 # See https://maven.apache.org/guides/mini/guide-naming-conventions.html
-mod_group_id=net.eroz33.runicrepository
+mod_group_id=net.eroz33.runic_repository
 # The authors of the mod. This is a simple text string that is used for display purposes in the mod list.
 mod_authors=Eroz33
 # The description of the mod. This is a simple multiline text string that is used for display purposes in the mod list.
