+++ src/main/java/net/eroz33/runic_repository/block/GrimoireBlock.java
@@ -1,0 +1,11 @@
+package net.eroz33.runic_repository.block;
+
+import net.minecraft.world.level.block.Block;
+
+
+public class GrimoireBlock extends Block {
+
+    public GrimoireBlock(Properties properties) {
+        super(properties);
+    }
+}
