+++ src/main/java/net/eroz33/runic_repository/block/TomeShelfBlock.java
@@ -1,0 +1,10 @@
+package net.eroz33.runic_repository.block;
+
+import net.minecraft.world.level.block.Block;
+
+
+public class TomeShelfBlock extends Block {
+    public TomeShelfBlock(Properties properties) {
+        super(properties);
+    }
+}
