+++ src/main/java/net/eroz33/runic_repository/datagen/RBlockTagProvider.java
@@ -1,0 +1,23 @@
+package net.eroz33.runic_repository.datagen;
+
+import net.minecraft.core.HolderLookup;
+import net.minecraft.data.PackOutput;
+import net.minecraft.tags.BlockTags;
+import net.neoforged.neoforge.common.data.BlockTagsProvider;
+import net.neoforged.neoforge.common.data.ExistingFileHelper;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.concurrent.CompletableFuture;
+
+import static net.eroz33.runic_repository.core.RR.MOD_ID;
+
+public class RBlockTagProvider extends BlockTagsProvider {
+    public RBlockTagProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider, @Nullable ExistingFileHelper existingFileHelper) {
+        super(output, lookupProvider, MOD_ID, existingFileHelper);
+    }
+
+    @Override
+    protected void addTags(HolderLookup.Provider provider) {
+
+    }
+}
