+++ src/main/java/net/eroz33/runic_repository/datagen/RItemTagProvider.java
@@ -1,0 +1,24 @@
+package net.eroz33.runic_repository.datagen;
+
+import net.minecraft.core.HolderLookup;
+import net.minecraft.data.PackOutput;
+import net.minecraft.data.tags.ItemTagsProvider;
+import net.minecraft.world.level.block.Block;
+import net.neoforged.neoforge.common.data.ExistingFileHelper;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.concurrent.CompletableFuture;
+
+import static net.eroz33.runic_repository.core.RR.MOD_ID;
+
+public class RItemTagProvider extends ItemTagsProvider {
+    public RItemTagProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider,
+                            CompletableFuture<TagLookup<Block>> blockTags, @Nullable ExistingFileHelper existingFileHelper) {
+        super(output, lookupProvider, blockTags, MOD_ID, existingFileHelper);
+    }
+
+    @Override
+    protected void addTags(HolderLookup.Provider provider) {
+
+    }
+}
