+++ src/main/java/net/eroz33/runic_repository/item/ArcaneDustItem.java
@@ -1,0 +1,10 @@
+package net.eroz33.runic_repository.item;
+
+import net.minecraft.world.item.Item;
+
+public class ArcaneDustItem extends Item {
+
+    public ArcaneDustItem() {
+        super (new Item.Properties());
+    }
+}
