+++ src/main/java/net/eroz33/runic_repository/item/ArcanePrimerItem.java
@@ -1,0 +1,9 @@
+package net.eroz33.runic_repository.item;
+
+import net.minecraft.world.item.Item;
+
+public class ArcanePrimerItem extends Item {
+    public ArcanePrimerItem() {
+        super (new Item.Properties());
+    }
+}
