--- src/main/java/net/eroz33/runicrepository/block/GrimoireBlock.java
@@ -1,11 +1,0 @@
-package net.eroz33.runicrepository.block;
-
-import net.minecraft.world.level.block.Block;
-
-
-public class GrimoireBlock extends Block {
-
-    public GrimoireBlock(Properties properties) {
-        super(properties);
-    }
-}
