--- src/main/java/net/eroz33/runicrepository/block/TomeShelfBlock.java
@@ -1,10 +1,0 @@
-package net.eroz33.runicrepository.block;
-
-import net.minecraft.world.level.block.Block;
-
-
-public class TomeShelfBlock extends Block {
-    public TomeShelfBlock(Properties properties) {
-        super(properties);
-    }
-}
