--- src/main/java/net/eroz33/runicrepository/core/RR.java
@@ -1,50 +1,0 @@
-package net.eroz33.runicrepository.core;
-
-import net.eroz33.runicrepository.config.ClientConfig;
-import net.eroz33.runicrepository.config.ServerConfig;
-import net.eroz33.runicrepository.setup.ClientSetup;
-import net.eroz33.runicrepository.setup.CommonSetup;
-import net.eroz33.runicrepository.setup.ServerSetup;
-
-import net.neoforged.api.distmarker.Dist;
-import net.neoforged.bus.api.IEventBus;
-import net.neoforged.fml.ModContainer;
-import net.neoforged.fml.common.Mod;
-import net.neoforged.fml.config.ModConfig;
-import net.neoforged.fml.loading.FMLEnvironment;
-import net.neoforged.neoforge.common.NeoForge;
-
-
-@Mod(RR.MOD_ID)
-public final class RR
-{
-    public static final String MOD_ID = "runicrepository";
-    public static final String NAME = "Runic Repository";
-
-    public static final ServerConfig SERVER_CONFIG = new ServerConfig();
-    public static final ClientConfig CLIENT_CONFIG = new ClientConfig();
-
-
-    public RR(IEventBus eventBus, ModContainer container)
-    {
-        // If we are on the client
-        if (FMLEnvironment.dist == Dist.CLIENT){
-            eventBus.addListener(ClientSetup::onClientSetup);
-            eventBus.addListener(ClientSetup::onRegisterKeymappings);
-        }
-
-        NeoForge.EVENT_BUS.register(new ServerSetup());
-
-        container.registerConfig(ModConfig.Type.SERVER, SERVER_CONFIG.getSpec());
-        container.registerConfig(ModConfig.Type.CLIENT, CLIENT_CONFIG.getSpec());
-
-        RBlocks.register(eventBus);
-        RBlockEntities.register(eventBus);
-        RItems.register(eventBus);
-
-        eventBus.addListener(CommonSetup::onCommonSetup);
-        eventBus.addListener(CommonSetup::onRegister);
-        eventBus.addListener(CommonSetup::onRegisterCapabilities);
-
-    }
-}
