--- src/main/java/net/eroz33/runicrepository/item/ArcaneDustItem.java
@@ -1,10 +1,0 @@
-package net.eroz33.runicrepository.item;
-
-import net.minecraft.world.item.Item;
-
-public class ArcaneDustItem extends Item {
-
-    public ArcaneDustItem() {
-        super (new Item.Properties());
-    }
-}
