--- src/main/java/net/eroz33/runicrepository/item/ArcanePrimerItem.java
@@ -1,9 +1,0 @@
-package net.eroz33.runicrepository.item;
-
-import net.minecraft.world.item.Item;
-
-public class ArcanePrimerItem extends Item {
-    public ArcanePrimerItem() {
-        super (new Item.Properties());
-    }
-}
