+++ src/main/resources/assets/runic_repository/textures/item/arcane_primer.png
