+++ src/main/resources/assets/runic_repository/textures/item/runic_core_g.png
