--- src/main/resources/assets/runicrepository/lang/en_us.json
@@ -1,15 +1,0 @@
-{
-  "item.runicrepository.arcane_dust": "Arcane Dust",
-  "item.runicrepository.arcane_primer": "Primer",
-
-  "item.runicrepository.runic_tome": "Runic Tome",
-  "item.runicrepository.arcane_tome": "Arcane Tome",
-  "item.runicrepository.wizened_tome": "Wizened Tome",
-  "item.runicrepository.mystic_tome": "Mystic Tome",
-
-
-  "block.runicrepository.grimoire": "Grimoire",
-  "block.runicrepository.tome_shelf": "Tome Shelf",
-
-  "creativetab.runicrepository": "Runic Repository"
-}
