--- src/main/resources/data/runicrepository/recipe/arcane_primer.json
@@ -1,21 +1,0 @@
-{
-  "type" : "minecraft:crafting_shaped",
-  "category": "misc",
-  "key": {
-    "B": {
-      "item" : "minecraft:book"
-    },
-    "D" : {
-      "item" : "runicrepository:arcane_dust"
-    }
-  },
-  "pattern" : [
-    "DDD",
-    "DBD",
-    "DDD"
-  ],
-  "result" : {
-    "count": 1,
-    "id": "runicrepository:arcane_primer"
-  }
-}
